// App/Models/FplBudgetLeagueInfo.ts
// Generated by https://quicktype.io

export interface FplBudgetLeagueInfo {
    new_entries:       NewEntries;
    last_updated_data: string;
    league:            League;
    standings:         NewEntries;
}

export interface League {
    id:           number;
    name:         string;
    created:      string;
    closed:       boolean;
    max_entries:  null;
    league_type:  string;
    scoring:      string;
    admin_entry:  null;
    start_event:  number;
    code_privacy: string;
    has_cup:      boolean;
    cup_league:   number;
    rank:         null;
}

export interface NewEntries {
    has_next: boolean;
    page:     number;
    results:  Result[];
}

export interface Result {
    id:          number;
    event_total: number;
    player_name: string;
    rank:        number;
    last_rank:   number;
    rank_sort:   number;
    total:       number;
    entry:       number;
    entry_name:  string;
}